fix: Correct Notification.ProgressStyle.Segment instantiation
Changed `Notification.ProgressStyle.Segment(segmentDuration)` to `Notification.ProgressStyle.Segment(segmentDuration.toInt())` to match the constructor signature which expects an Int.

+++ app/src/main/java/com/d4viddf/medicationreminder/services/PreReminderForegroundService.kt
@@ -202,7 +202,7 @@ class PreReminderForegroundService : Service() {
 
             for (i in 0 until segmentCount) {
                 segments.add(
-                    Notification.ProgressStyle.Segment(segmentDuration) // Use float for segment length
+                    Notification.ProgressStyle.Segment(segmentDuration.toInt()) // Corrected to use .toInt()
                         .setColor(colors[i % colors.size].toColorInt())
                 )
             }
